+++ packages/anthropic/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/anthropic
 
+## 4.0.24
+
+### Patch Changes
+
+- Updated dependencies [1659cd5]
+- Updated dependencies [6a5bdff]
+  - @ai-sdk/provider-utils@5.0.15
+
 ## 4.0.23
 
 ### Patch Changes