+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ai-sdk/azure
 
+## 4.0.25
+
+### Patch Changes
+
+- Updated dependencies [1659cd5]
+- Updated dependencies [6a5bdff]
+  - @ai-sdk/provider-utils@5.0.15
+  - @ai-sdk/deepseek@3.0.16
+  - @ai-sdk/openai@4.0.24
+
 ## 4.0.24
 
 ### Patch Changes